Update main.lua
Attempted to remove the loop, as it would be better to use the UpdateObject event if sql wasn't called each time the function ran.

So I fixed that issue.

Then discovered that the event isn't triggered from within the core.

I submitted a bug report on the issue. In the mean time the while loop will have to function as a work-a-round until someone gets around to fixing the event in the core.